% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion.R
\name{complete}
\alias{complete}
\title{Calculate differences in sea temperatures by month for different years and depths}
\usage{
complete(data, year, period, depth)
}
\arguments{
\item{data}{dataframe to analyze}

\item{year}{year to analyze}

\item{period}{period to analyze}

\item{depth}{depth to analyze}
}
\value{
A list with different information about the differences in temperature and average temperatures.
}
\description{
Calculate differences in sea temperatures by month for different years and depths
}
